refactor: switched Slot to generic class
